Create Alternating characters
You are given a string containing characters A  and B only. Your task is to change it into a string such that there are no matching adjacent characters. To do this, you are allowed to delete zero or more characters in the string.

Your task is to find the minimum number of required deletions.

For example, given the string s=AABAAB , remove an  at positions  and  to make s=ABAB in  deletions.