https://leetcode.com/problems/maximum-subarray/
https://www.youtube.com/watch?v=86CQq3pKSUw - Kadane's algorithm explanation

The logic is to find the subarray with largest sum in the array & the subarray should be contagious.

The largest value should be the value at the current index or adding the previous highest subarray with the current index value( that is curr+nums[i]). 

Then updating the highest value (nums[i] or curr). Then compare that value with the global highest subarray sum value and updating the highest value in the glob variable.